fix (webpack): don't include wp-polyfill dependencies automatically
Follow up to #278, remove option `injectPolyfill` in webpack plugin
dependency-extraction-webpack-plugin. The `wp-polyfill` script won't
be added as a dependency to all entries anymore. This will reduce the
number of scripts loaded on the page and save on resources.
Scripts will need to explicitely declare the dependencies if they
want to include it. The dependency will also be included if one of
the script dependency requires it.